requirements.txt
beautifulsoup4==4.9.3
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
et-xmlfile==1.1.0
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
pdfminer.six==20201018
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
xlrd==2.0.1
XlsxWriter==1.4.3
